backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "upstream-ecom"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=24.1",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "django-redis>=5.4",
    "gunicorn>=22.0",
    "stripe>=8.0",
    "whitenoise>=6.6",
    "python-json-logger>=2.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "factory-boy>=3.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 120
